+++ globe-landing/src/App.tsx
@@ -1,14 +1,8 @@
+import { Routes, Route } from 'react-router-dom';
 import { Navbar } from './components/Navbar';
-import { Hero } from './components/Hero';
-import { About } from './components/About';
-import { Ecosystem } from './components/Ecosystem';
-import { SiemensBanner } from './components/SiemensBanner';
-import { Differentiation } from './components/Differentiation';
-import { Stats } from './components/Stats';
-import { Presence } from './components/Presence';
-import { Innovation } from './components/Innovation';
-import { Contact } from './components/Contact';
 import { Footer } from './components/Footer';
+import { InicioPage } from './pages/InicioPage';
+import { GlobeServicesPage } from './pages/GlobeServicesPage';
 
 function App() {
   return (
@@ -16,15 +10,10 @@ function App() {
       <Navbar />
       {/* Spacer for fixed navbar */}
       <div className="h-[88px] lg:h-[116px]" />
-      <Hero />
-      <About />
-      <Ecosystem />
-      <SiemensBanner />
-      <Differentiation />
-      <Stats />
-      <Presence />
-      <Innovation />
-      <Contact />
+      <Routes>
+        <Route path="/" element={<InicioPage />} />
+        <Route path="/globe-services" element={<GlobeServicesPage />} />
+      </Routes>
       <Footer />
     </div>
   );